+++ src/OpenTelemetry.ResourceDetectors.Azure/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+* Added Azure VM resource detector.
+([#1182](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/1182))
+
 ## 1.0.0-alpha.1
 
 Released 2023-Apr-19